requirements.txt
django==3.0.6
djangorestframework==3.11.0
djangorestframework_jwt==1.11.0
python-decouple==3.1